+++ jre_lwjgl3glfw/src/main/java/org/lwjgl/opengl/Display.java
@@ -297,7 +297,7 @@ public class Display {
         Window.charCallback = new GLFWCharCallback() {
             @Override
             public void invoke(long window, int codepoint) {
-                GLFWInputImplementation.singleton.putKeyboardEvent(0,(byte)1,(int)codepoint,Sys.getNanoTime(),false);
+                GLFWInputImplementation.singleton.putKeyboardEvent(0,(byte)1, codepoint,Sys.getNanoTime(),false);
                 //GLFWInputImplementation.singleton.putKeyboardEvent(0,(byte)0,(int)codepoint,Sys.getNanoTime(),false);
             }
         };
